+++ src/senaite/ast/calc.py
@@ -27,6 +27,8 @@ from senaite.ast.config import RESISTANCE_KEY
 from senaite.ast.config import ZONE_SIZE_KEY
 from senaite.ast.utils import get_breakpoint
 from senaite.ast.utils import get_microorganism
+from senaite.ast.utils import get_sensitivity_category
+from senaite.ast.utils import get_sensitivity_category_value
 
 
 def calc_sensitivity_category(analysis_brain_uid, default_return='-'):
@@ -90,22 +92,13 @@ def calc_sensitivity_category(analysis_brain_uid, default_return='-'):
 
         # Get the breakpoint for this microorganism and antibiotic
         breakpoint = get_breakpoint(breakpoints_uid, microorganism, abx_uid)
-        if not breakpoint:
-            continue
 
-        # Choices for sensitivity category 0:|1:S|2:I|3:R
-        diameter_s = api.to_float(breakpoint.get("diameter_s"))
-        diameter_r = api.to_float(breakpoint.get("diameter_r"))
-        zone_size = api.to_float(zone_size)
-        if zone_size < diameter_r:
-            # R: resistant
-            category.update({"value": "3"})
-        elif zone_size >= diameter_s:
-            # S: sensible
-            category.update({"value": "1"})
-        else:
-            # I: susceptible at increased exposure
-            category.update({"value": "2"})
+        # Get the sensitivity category (S|I|R) and choice value
+        key = get_sensitivity_category(zone_size, breakpoint, default="")
+        value = get_sensitivity_category_value(key, default="")
+
+        # Update the sensitivity category
+        category.update({"value": value})
 
     # Assign the updated categories to the resistance analysis
     resistance_analysis.setInterimFields(categories)
@@ -116,6 +109,8 @@ def calc_sensitivity_category(analysis_brain_uid, default_return='-'):
         # Let's set the result as '-' so user can directly submit the whole
         # analysis without the need of confirming every single one
         resistance_analysis.setResult("-")
+    else:
+        resistance_analysis.setResult("")
 
     # Update disk dosage / concentration
     disk_dosages_analysis = get_by_keyword(analyses, DISK_CONTENT_KEY)