+++ docs/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- #12 Negative values for diameter and zone size tests not permitted
+- #12 Do not allow to submit AST analyses with empties
 - #11 Better styling of AST Panel selector in Sample view
 - #10 Allow to remove retracted AST analyses retests
 - Initial Release